feat: (WIP) Story map data layers
