+++ src/localization/locales/en-US.json
@@ -1016,6 +1016,11 @@
         "form_location_helper_text_step_3_image_src": "/storyMap/set-map-step-3-en.png",
         "form_location_helper_text_step_3_image_alt": "A screenshot of the Set Map Location screen. The “Set Location” button in the top right is circled.",
         "form_location_helper_text_bearing_icon_alt": "Bearing",
+        "form_location_add_data_layer_button": "Add Data Layer",
+        "form_location_add_data_layer_dialog_title": "<0>Add a data layer to </0><1>{{title}}</1>",
+        "form_location_add_data_layer_dialog_title_blank": "Add a data layer to chapter",
+        "form_location_add_data_layer_dialog_cancel": "Cancel",
+        "form_location_add_data_layer_confirm": "Add Map",
         "share_dialog_title": "Invite Editors",
         "share_dialog_description": "Invite people to edit your story map. <1>Learn more about collaborating on story maps</1>.",
         "share_dialog_help_url": "https://terraso.org/help/how-to-manage-editors-for-your-story-map/",
